compilation fixes for Integer -> Byte shift
Former-commit-id: c7e16c727b4c8c3fd79152956629e85bebade47c

+++ deeplearning4j-nlp-parent/deeplearning4j-nlp/src/main/java/org/deeplearning4j/models/word2vec/Huffman.java
@@ -66,8 +66,8 @@ public class Huffman {
     public void build() {
         buildTrigger = true;
         long[] count = new long[words.size() * 2 + 1];
-        int[] binary = new int[words.size() * 2 + 1];
-        int[] code = new int[MAX_CODE_LENGTH];
+        byte[] binary = new byte[words.size() * 2 + 1];
+        byte[] code = new byte[MAX_CODE_LENGTH];
         int[] point = new int[MAX_CODE_LENGTH];
         int[] parentNode = new int[words.size() * 2 + 1];
         int a = 0;
